fuzz: move `MockedDescriptorConverter` to `fuzz/util`
